接口优化更新

+++ front/src/main/java/com/cn/wanxi/front/cart/CartController.java
@@ -1,5 +1,6 @@
 package com.cn.wanxi.front.cart;
 
+import com.alibaba.druid.util.StringUtils;
 import com.cn.wanxi.model.cart.WxTabCart;
 import com.cn.wanxi.model.cart.WxTabSku;
 import com.cn.wanxi.service.cart.CartService;
@@ -11,6 +12,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 import java.util.Map;
+
+import static com.cn.wanxi.util.WebTools.returnData;
 
 
 /**
@@ -37,7 +40,9 @@ public class CartController {
 
     @RequestMapping(value = "/addCart",method = RequestMethod.POST)
     public Map<String,Object> addCart(WxTabCart wxTabCart){
+        if(StringUtils.isEmpty(wxTabCart.getNum()))return returnData("num不能为空",1);
 
+        if(StringUtils.isEmpty(wxTabCart.getNum()))return returnData("Spec不能为空",1);
         return cartService.addCart(wxTabCart);
     }
 
@@ -50,7 +55,9 @@ public class CartController {
      * @return
      */
     @RequestMapping("/updateNum")
-    public Map<String,Object> updateNum(@RequestParam(required = true) int id,@RequestParam(required = true) String num){
+    public Map<String,Object> updateNum(@RequestParam(required = true) Integer id,@RequestParam(required = true) String num){
+        if(StringUtils.isEmpty(String.valueOf(id)))return returnData("id不能为空",1);
+        if(StringUtils.isEmpty(num))return returnData("num不能为空",1);
 
         return cartService.updateNum(id,num);
     }
@@ -61,7 +68,8 @@ public class CartController {
      * @return
      */
     @RequestMapping("/deleteCart")
-    public Map<String,Object> deleteCart(@RequestParam(required = true) int id){
+    public Map<String,Object> deleteCart(@RequestParam(required = true) Integer id){
+        if(StringUtils.isEmpty(String.valueOf(id)))return returnData("id不能为空",1);
 
         return cartService.deleteCart(id);
     }
@@ -75,7 +83,7 @@ public class CartController {
      * @return
      */
     @RequestMapping(value = "/findCartList",method = RequestMethod.POST)
-    public List<Map<String,Object>> findCartList(@RequestParam(required = true) int page,@RequestParam(required = true) int size){
+    public List<Map<String,Object>> findCartList(@RequestParam(required = true) Integer page,@RequestParam(required = true) Integer size){
 
         return  cartService.findCartList(page,size);
     }
@@ -88,8 +96,7 @@ public class CartController {
      * @return
      */
     @RequestMapping(value ="/getSkuid",method = RequestMethod.POST)
-    public WxTabSku getSkuid(@RequestParam(required = true) int spuid,@RequestParam(required = true) String spec){
-
+    public WxTabSku getSkuid(@RequestParam(required = true) Integer spuid,@RequestParam(required = true) String spec){
         return cartService.getSkuid(spuid,spec);
     }
     /**
@@ -99,7 +106,8 @@ public class CartController {
      * @return
      */
     @RequestMapping(value ="/cardDetail",method = RequestMethod.POST)
-    public  Map<String,Object> cardDetail(@RequestParam(required = true) int id){
+    public  Map<String,Object> cardDetail( Integer id){
+        if(StringUtils.isEmpty(String.valueOf(id)))return returnData("id不能为空",1);
         return cartService.cardDetail(id);
     }
 }
